+++ examples/MicroserviceExample/Utils/Messaging/MessageSender.cs
@@ -28,7 +28,7 @@ namespace Utils.Messaging
     public class MessageSender : IDisposable
     {
         private static readonly ActivitySource ActivitySource = new ActivitySource(nameof(MessageSender));
-        private static readonly IPropagator Propagator = new TextMapPropagator();
+        private static readonly TextMapPropagator Propagator = Propagators.DefaultTextMapPropagator;
 
         private readonly ILogger<MessageSender> logger;
         private readonly IConnection connection;
@@ -59,15 +59,26 @@ namespace Utils.Messaging
                 {
                     var props = this.channel.CreateBasicProperties();
 
+                    // Depending on Sampling (and whether a listener is registered or not), the
+                    // activity above may not be created.
+                    // If it is created, then propagate its context.
+                    // If it is not created, the propagate the Current context,
+                    // if any.
+                    ActivityContext contextToInject = default;
                     if (activity != null)
                     {
-                        // Inject the ActivityContext into the message headers to propagate trace context to the receiving service.
-                        Propagator.Inject(new PropagationContext(activity.Context, Baggage.Current), props, this.InjectTraceContextIntoBasicProperties);
-
-                        // The OpenTelemetry messaging specification defines a number of attributes. These attributes are added here.
-                        RabbitMqHelper.AddMessagingTags(activity);
+                        contextToInject = activity.Context;
+                    }
+                    else if (Activity.Current != null)
+                    {
+                        contextToInject = Activity.Current.Context;
                     }
 
+                    // Inject the ActivityContext into the message headers to propagate trace context to the receiving service.
+                    Propagator.Inject(new PropagationContext(contextToInject, Baggage.Current), props, this.InjectTraceContextIntoBasicProperties);
+
+                    // The OpenTelemetry messaging specification defines a number of attributes. These attributes are added here.
+                    RabbitMqHelper.AddMessagingTags(activity);
                     var body = $"Published message: DateTime.Now = {DateTime.Now}.";
 
                     this.channel.BasicPublish(